# src/models/usermessages.ts
import mongoose from "mongoose";

const messagesSchema = new mongoose.Schema({
    name: {
        type: String,
        required: true
    },
    email: {
        type: String,
        required: true
    },
    message: {
        type: String,
        required: true
    },
    createdAt: {
        type: Date,
        default: Date.now
    }
});


// Create and export Message model
const Message = mongoose.model('Message', messagesSchema);
export default Message;